% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{api}
\alias{api}
\title{API: Make a URL to download data from BCCR}
\usage{
api(cuadro, first = 2012, last = 2015)
}
\arguments{
\item{cuadro}{A number identifying the BCCR's data table (integer).}

\item{first}{The first year to download (integer, default=1990).}

\item{last}{The last year to download (integer, default=2015).}
}
\value{
A valid URL to download the data
}
\description{
\code{api} returns the URL for a BCCR table, for a given year range.
}
\examples{
api(138, 1995, 2016)
}

