% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{daily_to_monthly}
\alias{daily_to_monthly}
\title{Convert daily data to monthly}
\usage{
daily_to_monthly(datos, func)
}
\arguments{
\item{datos}{A data.table where dates are indicated by a lubridate vector "fecha"}

\item{func}{A function to summarize the data}
}
\value{
A data.table with monthly data
}
\description{
Convert daily data to monthly
}
\examples{
ss <- read_daily_series(list(tbasica=17,tc=367), 2010, 2015)
daily_to_monthly(ss, dplyr::last)
daily_to_monthly(ss, mean)
}

