% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{download_series}
\alias{download_series}
\title{Download a table from the BCCR website.}
\usage{
download_series(cuadro, first = 2012, last = 2015)
}
\arguments{
\item{cuadro}{A number identifying the BCCR's data table (integer).}

\item{first}{The first year to download (integer, default=1990).}

\item{last}{The last year to download (integer, default=2015).}
}
\value{
A data.table with data in same format as in BCCR website.
}
\description{
\code{download_series} downloads data from the BCCR website, stripping header rows
}
\examples{
download_series(125)
download_series(138, 1995, 2016)
download_series(367, header=5)
}

