% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{find_series}
\alias{find_series}
\title{Find indicators by name}
\usage{
find_series(name)
}
\arguments{
\item{name}{A string text with part of the indicator name}
}
\value{
A data.frame with indicators whose names match the requested value
}
\description{
Find indicators by name
}
\examples{
find_series("ipc")
find_series("agricultura")
}

