% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fix.spanish.chars}
\alias{fix.spanish.chars}
\title{Fix wrong Spanish characters}
\usage{
fix.spanish.chars(txt)
}
\arguments{
\item{txt}{(Vector of) strings with wrong characters}
}
\value{
txt (vector of) strings with proper Spanish characters
}
\description{
Fix wrong Spanish characters
}

