% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mgsub}
\alias{mgsub}
\title{Multiple replacement}
\usage{
mgsub(pattern, replacement, x, ...)
}
\arguments{
\item{pattern}{A vector of strings to be replaced}

\item{replacement}{A matching vector of strings with the replacements}

\item{x}{Text to be modified}

\item{...}{Additional parameters to be passed to gsub}
}
\value{
The modified text
}
\description{
Multiple replacement
}
\examples{
mgsub(wrong, right, text)
}

