% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{read_daily_series}
\alias{read_daily_series}
\title{Reads daily series.}
\usage{
read_daily_series(series, first = 1950, last = lubridate::year(Sys.Date()),
  freq = "d", func = mean)
}
\arguments{
\item{series}{A list or a table. If a list is provided, its entries should be
  "cuadro" numbers, labeled with the desired variable name. If a table is
  provided, its first column must have the desired variable names, and its
  second column the \emph{cuadro} numbers.

  In all case, \emph{cuadro} is the number of a table in the BCCR website.}

\item{first}{The first year to download (integer, default=1990).}

\item{last}{The last year to download (integer, default=2015).}
}
\value{
A data.table with given series.
}
\description{
\code{read_daily_series} is used to download tables of a daily series, where
each row is a day of the year and each column a year.
}
\examples{
mylist <- list(tc=367, tbasica=17)
dd <- read_daily_series(mylist)
}

