% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{read_indicator_year}
\alias{read_indicator_year}
\title{Title}
\usage{
read_indicator_year(cuadro, first = 1950,
  last = lubridate::year(Sys.Date()))
}
\arguments{
\item{last}{}
}
\description{
Title
}

