% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{read_month_year}
\alias{read_month_year}
\title{Reads monthly series.}
\usage{
read_month_year(series, first = 1950, last = lubridate::year(Sys.Date()))
}
\arguments{
\item{series}{A list of name=cuadro pairs. Name (a string) is the name of a
series, cuadro (integer) the number of table in the BCCR website.}

\item{first}{The first year to download (integer, default=1990).}

\item{last}{The last year to download (integer, default=2015).}
}
\value{
A data frame with given series.
}
\description{
\code{read_montly_series} is used to download tables of montly data series,
where each row is a month and each column a year.
}
\examples{
read_month_year(list(M1=125))
}

