% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{read_monthly_series}
\alias{read_monthly_series}
\title{Reads montly series.}
\usage{
read_monthly_series(series, first = 1990, last = 2015, header = 5)
}
\arguments{
\item{series}{A list of name=cuadro pairs. Name (a string) is the name of a
series, cuadro (integer) the number of table in the BCCR website.}

\item{first}{The first year to download (integer, default=1990).}

\item{last}{The last year to download (integer, default=2015).}

\item{header}{Number of the line that contains the column headers (integer, default=6)}
}
\value{
A data frame with given series.
}
\description{
\code{read_montly_series} is used to download tables of montly data series,
where each row is a year and each column a month.
}
\examples{
read_montly_series('M1', 125)
}

