% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{read_titles}
\alias{read_titles}
\title{Title}
\usage{
read_titles(series)
}
\arguments{
\item{series}{A vector of table numbers}
}
\value{
A table
}
\description{
Extracts title information (first two lines) for given table numbers
}
\examples{
read_titles(c(125, 96))
}

