% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{read_year_month}
\alias{read_year_month}
\title{Reads monthly series.}
\usage{
read_year_month(series, first = 1950, last = lubridate::year(Sys.Date()))
}
\arguments{
\item{series}{A list of name=cuadro pairs. Name (a string) is the name of a
series, cuadro (integer) the number of table in the BCCR website.}

\item{first}{The first year to download (integer, default=1990).}

\item{last}{The last year to download (integer, default=2015).}
}
\value{
A data frame with given series.
}
\description{
\code{read_year_month} is used to download tables of montly data series,
where each row is a year and each column a month.
}
\examples{
read_year_month(list(lmn=95, lme=96))
}

