% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{remove_empty_columns}
\alias{remove_empty_columns}
\title{Remove empty columns}
\usage{
remove_empty_columns(datos)
}
\arguments{
\item{datos}{A data.frame}
}
\value{
Same data.frame, after removing columns with no data
}
\description{
Remove empty columns
}
\examples{
remove_empty_columns(mydata)
}

