% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{subs_commas}
\alias{subs_commas}
\title{Fix the decimal separator}
\usage{
subs_commas(value)
}
\arguments{
\item{value}{A vector of numbers (with comma decimals)}
}
\value{
The same vector of numbers, but with period decimals
}
\description{
In Costa Rica, the decimal separator is indicated by a comma (,) while R
expects a period (.). \code{subs_commas} replaces commas by periods.
}
\examples{
subs_commas(3,14) # -> 3.14
}

