% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{table_to_list}
\alias{table_to_list}
\title{Table to list}
\usage{
table_to_list(tab)
}
\arguments{
\item{tab}{A table, whose first column is used as labels (variable names) and second column as values ("cuadro" number)}
}
\value{
A list
}
\description{
Converts a table to a (dictionary) list. Labels are taken from first column and values from second column.
Used by \code{read_daily_series} to accept table input instead of lists.
}

