% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{trim_dataframe}
\alias{trim_dataframe}
\title{Trim a data frame}
\usage{
trim_dataframe(df)
}
\arguments{
\item{df}{A data.table}
}
\value{
A data.table
}
\description{
Removes initial and last observations where all variables have missing values.
}

