% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{YMD}
\alias{YMD}
\title{Makes a lubridate date}
\usage{
YMD(y, m, d = 0)
}
\arguments{
\item{y}{Year (integer).}

\item{m}{Month (integer).}

\item{d}{Day (integer). If zero (default), set d to last day of month}
}
\value{
A lubridate date
}
\description{
\code{YMD} returns the last date of the month, for given year and month.
}
\examples{
YMD(2010, 4)  # -> April 30th, 2010
YMD(2016, 2)  # -> February 29th, 2016
}

